+++ docs/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Releases
 
+## v1.24.0
+
+Released: ???
+
+Features:
+
+
 ## v1.23.0
+
+Released: 2025-01-19
 
 Features: